% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors_names_csv.R
\name{authors_names_csv}
\alias{authors_names_csv}
\title{Create a list of authors}
\usage{
authors_names_csv(file_path)
}
\arguments{
\item{file_path}{The full path of the CSV file}
}
\value{
A list of author/affiliation to be used with coverpage()
}
\description{
This function creates a list of authors/affiliations from a CSV file. The CSV should contain one authors per line, with the first column with the author full name, the subsequent for the affiliations (one affiliation per column only). The first column should be named "author".
}
