% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors_names_interactive.R
\name{authors_names_interactive}
\alias{authors_names_interactive}
\title{Create a list of authors}
\usage{
authors_names_interactive()
}
\value{
A list of author/affiliation to be used with coverpage()
}
\description{
This function creates a list of authors/affiliations interactively
}
