% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontpage.R
\name{frontpage}
\alias{frontpage}
\title{Scientfic article coverpage builder}
\usage{
frontpage(author_list, path)
}
\arguments{
\item{author_list}{A named list of author/affiliations, with the author names as names of the 
list, and affiliations as the content of the list. Can be created with the function 
author_names_csv()}

\item{path}{The full path of the coverpage ('path/filename.docx')}
}
\value{
Save a coverpage.docx with names and affiliations
}
\description{
This function creates a proper frontpage for scientific articles, with superscript numbers 
refering to authors affiliations
}
